+++ srcs/test.c
@@ -6,7 +6,7 @@
 /*   By: seokchoi <seokchoi@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/22 20:45:48 by seokchoi          #+#    #+#             */
-/*   Updated: 2022/11/28 16:26:25 by seokchoi         ###   ########.fr       */
+/*   Updated: 2022/11/28 21:18:17 by seokchoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -120,11 +120,11 @@ void	viewtree(t_token *tok)
 void	printf_key(t_keys *keys)
 {
 	t_keys			*keys_tmp;
- 
+
 	keys_tmp = keys; 
 	while (keys_tmp) // key 확인
 	{
-		printf("key = %s , value = %s\n", keys_tmp->key, keys_tmp->value);
+		printf("key = %s , value = '%s'\n", keys_tmp->key, keys_tmp->value);
 		keys_tmp = keys_tmp->next;
 	}
 }
